+++ src/edu/jhuapl/saavtk/pick/EllipsePicker.java
@@ -136,7 +136,7 @@ public class EllipsePicker extends Picker
 		vtkActor pickedActor = vStructureCP.GetActor();
 		if (pickedActor == refStructureManager.getBoundaryActor())
 		{
-			int cellId = vStructureCP.GetCellId();
+			int cellId = (int)vStructureCP.GetCellId();
 			int pointId = refStructureManager.getPolygonIdFromBoundaryCellId(cellId);
 			currVertexId = pointId;
 		}
